refactor route closures to controllers
